added spec to twitter client
