+++ social_panel/Gemfile.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+source 'https://rubygems.org'
+
+gem "faraday", "~> 1.0"
+
+gem "rspec", "~> 3.9"
+
+gem "json", "~> 2.3"
